requirements.txt
firebase-admin==7.2.0
loguru==0.7.3
python-dotenv==1.2.1
python-telegram-bot==22.5
